Move table and config to seperate file
