+++ packages/extension/src/view/devtools/components/privateAdvertising/protectedAudience/interestGroups/index.tsx
@@ -17,55 +17,16 @@
 /**
  * External dependencies.
  */
-import React, { useEffect, useMemo, useState } from 'react';
-import { Resizable } from 're-resizable';
-import {
-  Table,
-  type TableFilter,
-  TableProvider,
-  type TableColumn,
-  type TableRow,
-  useSidebar,
-  SIDEBAR_ITEMS_KEYS,
-  type TableData,
-} from '@google-psat/design-system';
-import { I18n } from '@google-psat/i18n';
-import {
-  noop,
-  type InterestGroups as InterestGroupsType,
-} from '@google-psat/common';
-import { prettyPrintJson } from 'pretty-print-json';
+import React from 'react';
+import { useSidebar, SIDEBAR_ITEMS_KEYS } from '@google-psat/design-system';
 
 /**
  * Internal dependencies.
  */
 import { useProtectedAudience, useSettings } from '../../../../stateProviders';
+import IGTable from './table';
 
-interface InterestGroupsProps {
-  filters?: Record<string, any>;
-}
-
-const InterestGroups = ({ filters }: InterestGroupsProps) => {
-  const [, setQuery] = useState({
-    filter: {},
-  });
-
-  useEffect(() => {
-    setQuery({
-      filter: {
-        ...filters,
-      },
-    });
-  }, [filters]);
-
-  // const clearQuery = useCallback(() => {
-  //   setQuery({
-  //     filter: {},
-  //   });
-  // }, []);
-
-  const [selectedRow, setSelectedRow] = useState<TableData | null>(null);
-
+const InterestGroups = () => {
   const { interestGroupDetails } = useProtectedAudience(({ state }) => ({
     interestGroupDetails: state.interestGroupDetails,
   }));
@@ -77,74 +38,6 @@ const InterestGroups = ({ filters }: InterestGroupsProps) => {
   const { updateSelectedItemKey } = useSidebar(({ actions }) => ({
     updateSelectedItemKey: actions.updateSelectedItemKey,
   }));
-
-  const tableColumns = useMemo<TableColumn[]>(
-    () => [
-      {
-        header: 'Event Time',
-        accessorKey: 'formattedTime',
-        cell: (info) =>
-          (info as string)
-            .replace('T', ' | ')
-            .replace('Z', '')
-            .split('-')
-            .join('/'),
-        enableHiding: false,
-      },
-      {
-        header: 'Access Type',
-        accessorKey: 'type',
-        cell: (info) => info,
-      },
-      {
-        header: 'Name',
-        accessorKey: 'name',
-        cell: (info) => info,
-      },
-      {
-        header: 'Owner',
-        accessorKey: 'ownerOrigin',
-        cell: (info) => {
-          return new URL(info as string).hostname;
-        },
-      },
-      {
-        header: 'Expiration Time',
-        accessorKey: 'details.expirationTime',
-        cell: (info) => {
-          if (!info) {
-            return '';
-          }
-
-          return new Date(Number(info) * 1000)
-            .toISOString()
-            .replace('T', ' | ')
-            .replace('Z', '')
-            .split('-')
-            .join('/');
-        },
-      },
-    ],
-    []
-  );
-
-  const tableFilters = useMemo<TableFilter>(
-    () => ({
-      type: {
-        title: 'Access Type',
-        sortValues: true,
-      },
-      ownerOrigin: {
-        title: 'Owner',
-        sortValues: true,
-      },
-      name: {
-        title: 'Name',
-        sortValues: true,
-      },
-    }),
-    []
-  );
 
   if (!isUsingCDP) {
     return (
@@ -178,67 +71,7 @@ const InterestGroups = ({ filters }: InterestGroupsProps) => {
     );
   }
 
-  return (
-    <div className="w-full h-full text-outer-space-crayola border-x border-american-silver dark:border-quartz flex flex-col">
-      <Resizable
-        defaultSize={{
-          width: '100%',
-          height: '80%',
-        }}
-        minHeight="15%"
-        maxHeight="90%"
-        enable={{
-          bottom: true,
-        }}
-      >
-        <TableProvider
-          data={interestGroupDetails}
-          tableColumns={tableColumns}
-          tableFilterData={tableFilters}
-          tableSearchKeys={undefined}
-          tablePersistentSettingsKey="interestGroupsTable"
-          onRowClick={(row) => {
-            setSelectedRow(row as InterestGroupsType);
-          }}
-          onRowContextMenu={noop}
-          getRowObjectKey={(row: TableRow) => {
-            return (
-              ((row.originalData as InterestGroupsType).uniqueAuctionId ?? '') +
-              (row.originalData as InterestGroupsType).time
-            );
-          }}
-        >
-          <Table
-            selectedKey={
-              (selectedRow?.uniqueAuctionId ?? '') + String(selectedRow?.time)
-            }
-            hideSearch={true}
-            minWidth="50rem"
-          />
-        </TableProvider>
-      </Resizable>
-      <div className="flex-1 text-raisin-black dark:text-bright-gray border border-gray-300 dark:border-quartz shadow h-full min-w-[10rem] bg-white dark:bg-raisin-black overflow-auto">
-        {selectedRow ? (
-          <div className="text-xs py-1 px-1.5">
-            <pre>
-              <div
-                className="json-container"
-                dangerouslySetInnerHTML={{
-                  __html: prettyPrintJson.toHtml(selectedRow),
-                }}
-              />
-            </pre>
-          </div>
-        ) : (
-          <div className="h-full p-8 flex items-center">
-            <p className="text-lg w-full font-bold text-granite-gray dark:text-manatee text-center">
-              {I18n.getMessage('selectRowToPreview')}
-            </p>
-          </div>
-        )}
-      </div>
-    </div>
-  );
+  return <IGTable interestGroupDetails={interestGroupDetails} />;
 };
 
 export default InterestGroups;
